Remove legacy extension store and mainsearch code.
